logstash/requirements.txt
-i https://pypi.org/simple
certifi==2023.11.17
charset-normalizer==3.3.2
datetime==5.4
idna==3.6
logdissect==3.1.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3 ; python_version >= '3.8'
urllib3==2.1.0
zope.interface==6.1
